// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  RIDE_OFFER = 'RIDE_OFFER',
  RIDE_ASSIGNED = 'RIDE_ASSIGNED',
  DRIVER_ARRIVING = 'DRIVER_ARRIVING',
  DRIVER_ARRIVED = 'DRIVER_ARRIVED',
  RIDE_STARTED = 'RIDE_STARTED',
  RIDE_COMPLETED = 'RIDE_COMPLETED',
  RIDE_CANCELLED = 'RIDE_CANCELLED',
  RIDE_EXPIRED = 'RIDE_EXPIRED',
  RATING_REQUEST = 'RATING_REQUEST',
}